# openapi.yaml
openapi: 3.0.3
info:
  title: ShopKart Capstone API
  version: 1.0.0
  description: Deterministic retail API supplied for the Week 6 pre-capstone assessment.
servers:
  - url: http://localhost:8080/api
tags:
  - name: Health
  - name: Authentication
  - name: Products
  - name: Carts
  - name: Orders
paths:
  /health:
    get:
      tags: [Health]
      summary: Verify the API and MySQL connection
      responses:
        "200":
          description: Service is ready
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Health"
  /auth/login:
    post:
      tags: [Authentication]
      summary: Exchange a seeded test-account login for a bearer token
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/LoginRequest"
      responses:
        "200":
          description: Authenticated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/LoginResponse"
        "401":
          $ref: "#/components/responses/Unauthorized"
  /products:
    get:
      tags: [Products]
      summary: Search the seeded product catalogue
      parameters:
        - in: query
          name: q
          schema:
            type: string
            maxLength: 80
      responses:
        "200":
          description: Matching products
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Product"
  /products/{sku}:
    get:
      tags: [Products]
      summary: Read one product
      parameters:
        - $ref: "#/components/parameters/Sku"
      responses:
        "200":
          description: Product found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Product"
        "404":
          $ref: "#/components/responses/NotFound"
  /carts:
    post:
      tags: [Carts]
      summary: Create an open cart for the authenticated customer
      security:
        - bearerAuth: []
      responses:
        "201":
          description: Cart created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Cart"
        "401":
          $ref: "#/components/responses/Unauthorized"
  /carts/{id}:
    get:
      tags: [Carts]
      summary: Read an owned cart
      security:
        - bearerAuth: []
      parameters:
        - $ref: "#/components/parameters/Id"
      responses:
        "200":
          description: Owned cart
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Cart"
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          $ref: "#/components/responses/NotFound"
  /carts/{id}/items:
    post:
      tags: [Carts]
      summary: Add a quantity of one SKU to an open cart
      security:
        - bearerAuth: []
      parameters:
        - $ref: "#/components/parameters/Id"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AddCartItemRequest"
      responses:
        "200":
          description: Updated cart and paise-safe total
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Cart"
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          $ref: "#/components/responses/NotFound"
        "409":
          description: Out of stock or cart is not open
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /orders:
    post:
      tags: [Orders]
      summary: Convert one owned, non-empty cart into exactly one order
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateOrderRequest"
      responses:
        "201":
          description: Order placed
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Order"
        "403":
          $ref: "#/components/responses/Forbidden"
        "409":
          description: Empty or previously ordered cart
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /orders/{id}:
    get:
      tags: [Orders]
      summary: Read an owned order
      security:
        - bearerAuth: []
      parameters:
        - $ref: "#/components/parameters/Id"
      responses:
        "200":
          description: Owned order
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Order"
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          $ref: "#/components/responses/NotFound"
  /orders/{id}/cancel:
    post:
      tags: [Orders]
      summary: Cancel a placed order
      security:
        - bearerAuth: []
      parameters:
        - $ref: "#/components/parameters/Id"
      responses:
        "200":
          description: Order cancelled
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Order"
        "403":
          $ref: "#/components/responses/Forbidden"
        "409":
          description: Order is no longer PLACED
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: ShopKart-HMAC
  parameters:
    Id:
      in: path
      name: id
      required: true
      schema:
        type: integer
        format: int64
        minimum: 1
    Sku:
      in: path
      name: sku
      required: true
      schema:
        type: string
        example: SKU-BAG
  responses:
    Unauthorized:
      description: Missing, invalid, expired, or incorrect authentication
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
    Forbidden:
      description: The object belongs to another authenticated customer
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
    NotFound:
      description: Resource not found
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
  schemas:
    Health:
      type: object
      required: [status, service, database]
      properties:
        status: { type: string, example: UP }
        service: { type: string, example: shopkart }
        database: { type: string, example: mysql }
    LoginRequest:
      type: object
      required: [email, password]
      properties:
        email: { type: string, format: email, example: alice@shopkart.test }
        password: { type: string, format: password, writeOnly: true }
    LoginResponse:
      type: object
      required: [token, customerId, customer]
      properties:
        token: { type: string, writeOnly: true }
        customerId: { type: integer, format: int64 }
        customer:
          $ref: "#/components/schemas/Customer"
    Customer:
      type: object
      required: [id, persona, email, displayName]
      properties:
        id: { type: integer, format: int64 }
        persona: { type: string, example: alice }
        email: { type: string, format: email }
        displayName: { type: string }
    Product:
      type: object
      required: [sku, name, description, category, pricePaise, stock, imageKey]
      properties:
        sku: { type: string, example: SKU-BAG }
        name: { type: string, example: Metro Carryall }
        description: { type: string }
        category: { type: string, example: Bags }
        pricePaise: { type: integer, minimum: 1, example: 49900 }
        stock: { type: integer, minimum: 0, example: 10 }
        imageKey: { type: string, example: bag }
    CartItem:
      type: object
      required: [sku, name, qty, unitPricePaise, lineTotalPaise]
      properties:
        sku: { type: string }
        name: { type: string }
        qty: { type: integer, minimum: 1 }
        unitPricePaise: { type: integer, minimum: 1 }
        lineTotalPaise: { type: integer, minimum: 1 }
    Cart:
      type: object
      required: [id, cartId, status, items, totalPaise]
      properties:
        id: { type: integer, format: int64 }
        cartId: { type: integer, format: int64 }
        status: { type: string, enum: [OPEN, ORDERED] }
        items:
          type: array
          items:
            $ref: "#/components/schemas/CartItem"
        totalPaise: { type: integer, minimum: 0 }
    AddCartItemRequest:
      type: object
      required: [sku, qty]
      properties:
        sku: { type: string, example: SKU-BAG }
        qty: { type: integer, minimum: 1, example: 2 }
    CreateOrderRequest:
      type: object
      required: [cartId, address]
      properties:
        cartId: { type: integer, format: int64, minimum: 1 }
        address: { type: string, minLength: 10, maxLength: 240 }
    Order:
      type: object
      required: [id, orderId, customerId, cartId, status, totalPaise, address, createdAt, items]
      properties:
        id: { type: integer, format: int64 }
        orderId: { type: integer, format: int64 }
        customerId: { type: integer, format: int64 }
        cartId: { type: integer, format: int64 }
        status: { type: string, enum: [PLACED, CANCELLED] }
        totalPaise: { type: integer, minimum: 1 }
        address: { type: string }
        createdAt: { type: string, format: date-time }
        items:
          type: array
          items:
            $ref: "#/components/schemas/CartItem"
    ErrorResponse:
      type: object
      required: [error]
      properties:
        error:
          type: object
          required: [code, message]
          properties:
            code: { type: string, example: ORDER_FORBIDDEN }
            message: { type: string }